+++ apis/common/common_types.go
@@ -6,3 +6,8 @@ type StandQueryParameters struct {
 	Page          int      `json:"page"`
 	PerPage       int      `json:"perPage"`
 }
+
+type BasicOperationResponse struct {
+	Code int    `json:"code"`
+	Data string `json:"data"`
+}